+++ src/index.js
@@ -1,4 +1,3 @@
-export * from './layer_/index'
-export * from './layer_a__/index.js'
-export * from './top_layer__/index.js'
-export * from './top_layer_zIndex__/index.js'
+export * from './layer/index.js'
+export * from './top_layer/index.js'
+export * from './top_layer_zIndex/index.js'