+++ src/layer_/index.d.ts
@@ -1,7 +1,0 @@
-/**
- * Returns a new layer.
- */
-export declare function layer_<Obj>(obj?:Obj):Obj&layer_T
-export interface layer_T {
-	zIndex:number|null
-}