+++ src/top_layer/index.d.ts
@@ -1,0 +1,7 @@
+import type { val__be_computed_pair_T } from '@ctx-core/nanostores'
+import type { Layer } from '../layer_a'
+export declare const [
+	top_layer$_,
+	top_layer_,
+]:val__be_computed_pair_T<Layer|undefined>
+export { top_layer$_ as top_layer__ }