+++ src/top_layer__/index.d.ts
@@ -1,5 +1,0 @@
-import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-import type { Layer } from '../layer_a__'
-export declare const top_layer__:Be<ReadableAtom_<Layer|undefined>>
-export { top_layer__ as top_layer$_ }