+++ src/top_layer_zIndex/index.d.ts
@@ -1,0 +1,6 @@
+import type { val__be_computed_pair_T } from '@ctx-core/nanostores'
+export declare const [
+	top_layer_zIndex$_,
+	top_layer_zIndex_,
+]:val__be_computed_pair_T<number|null>
+export { top_layer_zIndex$_ as top_layer_zIndex__ }