+++ src/top_layer_zIndex__/index.d.ts
@@ -1,4 +1,0 @@
-import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const top_layer_zIndex__:Be<ReadableAtom_<number|null>>
-export { top_layer_zIndex__ as top_layer_zIndex$_ }