+++ src/top_layer_zIndex__/index.js
@@ -1,8 +1,0 @@
-import { computed_ } from '@ctx-core/nanostores'
-import { be_ } from '@ctx-core/object'
-import { top_layer__ } from '../top_layer__/index.js'
-export const top_layer_zIndex__ = be_('top_layer_zIndex__', ctx=>
-	computed_(top_layer__(ctx),
-		top_layer=>
-			top_layer ? top_layer.zIndex : null))
-export { top_layer_zIndex__ as top_layer_zIndex$_ }